refactor(sitemap): extract config from sitemap parser
Moved SitemapConfig + SitemapConfigBuilder to sitemap_config.rs (148 lines).
sitemap_parser.rs: 753 → 581 lines (-23%).
Re-exports maintained via crawler/mod.rs for backward compat.
21/21 tests passing.
